+++ src/index.html
@@ -11,8 +11,8 @@
 <body>
   <!--ESTE ES UN COMENTARIO PARA VER SI SE PUEDE EL GIT PULL -->
     <section id="root"></section>
-<script type="module" src="main.js"></script>
+<script type="module" src="router.js"></script>
 <script type="module" src="./lib/index.js"></script>
-<script type="module" src="mainfinal.js"></script>
+<script type="module" src="main.js"></script>
 </body>
 </html>
