+++ src/lib/index.js
@@ -5,7 +5,7 @@ export const myFunction = () => {
   console.log('Hola mundo!');
 }; */
 // Import the functions you need from the SDKs you need
-import { onNavigate } from '../main.js';
+import { onNavigate } from '../router.js';
 import { initializeApp } from 'https://www.gstatic.com/firebasejs/9.6.8/firebase-app.js';
 import {
   getAuth,
